Address controller added

+++ Ma.rt/Mart/Controllers/AddressesController.cs
@@ -10,25 +10,34 @@ using System.Threading.Tasks;
 using System.Web.Http;
 using System.Web.Http.Description;
 using EF.mart;
+using Microsoft.AspNet.Identity;
 
 namespace Mart.Controllers
 {
+    [Authorize]
+    [RoutePrefix("api/Address")]
     public class AddressesController : ApiController
     {
         private cartDBEntitiesConn db = new cartDBEntitiesConn();
 
         // GET: api/Addresses
-        public IQueryable<Address> GetAddresses()
-        {
-            return db.Addresses;
-        }
+        //public IQueryable<Address> GetAddresses()
+        //{
+        //    return db.Addresses;
+        //}
 
         // GET: api/Addresses/5
+        [Route("GetAddress/{id:int}")]
         [ResponseType(typeof(Address))]
         public async Task<IHttpActionResult> GetAddress(int id)
         {
+            string userId = User.Identity.GetUserId();
+            db.Configuration.ProxyCreationEnabled = false;
+            Customer customer = db.Customers.FirstOrDefault(u => u.AspNetUserId == userId);
+
             Address address = await db.Addresses.FindAsync(id);
-            if (address == null)
+
+            if (address == null || address.CustomerUserName != customer.UserName)
             {
                 return NotFound();
             }
@@ -37,6 +46,7 @@ namespace Mart.Controllers
         }
 
         // PUT: api/Addresses/5
+        [Route("UpdateAddress/{id:int}")]
         [ResponseType(typeof(void))]
         public async Task<IHttpActionResult> PutAddress(int id, Address address)
         {
@@ -45,7 +55,11 @@ namespace Mart.Controllers
                 return BadRequest(ModelState);
             }
 
-            if (id != address.AddressId)
+            string userId = User.Identity.GetUserId();
+            db.Configuration.ProxyCreationEnabled = false;
+            Customer customer = db.Customers.FirstOrDefault(u => u.AspNetUserId == userId);
+
+            if (id != address.AddressId || address.CustomerUserName != customer.AspNetUserId)
             {
                 return BadRequest();
             }
@@ -72,12 +86,21 @@ namespace Mart.Controllers
         }
 
         // POST: api/Addresses
+        [Route("PostAddress/{id:int}")]
         [ResponseType(typeof(Address))]
         public async Task<IHttpActionResult> PostAddress(Address address)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
+            }
+            string userId = User.Identity.GetUserId();
+            db.Configuration.ProxyCreationEnabled = false;
+            Customer customer = db.Customers.FirstOrDefault(u => u.AspNetUserId == userId);
+
+            if (address.CustomerUserName != customer.AspNetUserId)
+            {
+                return BadRequest();
             }
 
             db.Addresses.Add(address);
@@ -87,6 +110,7 @@ namespace Mart.Controllers
         }
 
         // DELETE: api/Addresses/5
+        [Route("RemoveAddress/{id:int}")]
         [ResponseType(typeof(Address))]
         public async Task<IHttpActionResult> DeleteAddress(int id)
         {
@@ -95,6 +119,16 @@ namespace Mart.Controllers
             {
                 return NotFound();
             }
+
+            string userId = User.Identity.GetUserId();
+            db.Configuration.ProxyCreationEnabled = false;
+            Customer customer = db.Customers.FirstOrDefault(u => u.AspNetUserId == userId);
+
+            if (id != address.AddressId || address.CustomerUserName != customer.AspNetUserId)
+            {
+                return BadRequest();
+            }
+
 
             db.Addresses.Remove(address);
             await db.SaveChangesAsync();
